refactor: ユースケース層を分離し、クリーンアーキテクチャに準拠した構造へ変更
